# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Tec-Doc API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/",
    "paths": {
        "/api/v1/articles/enrichment": {
            "post": {
                "description": "to enrichment product by brand and article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetTecDocArticles",
                "operationId": "articles_enrichment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit of contents",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "offset of contents",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The input body",
                        "name": "InputBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetTecDocArticlesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Article"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errinfo.errInf"
                        }
                    }
                }
            }
        },
        "/api/v1/excel": {
            "get": {
                "description": "download excel table template",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet.main+xml"
                ],
                "tags": [
                    "excel"
                ],
                "summary": "ExcelTemplate",
                "operationId": "excel_template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errinfo.errInf"
                        }
                    }
                }
            },
            "post": {
                "description": "upload excel table containing products info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excel"
                ],
                "summary": "LoadFromExcel",
                "operationId": "load_from_excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "excel file",
                        "name": "excel_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/excel/enrichment": {
            "post": {
                "description": "Enrichment excel file, limit entiies in file = 10000",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "excel"
                ],
                "summary": "ProductsEnrichedExcel",
                "operationId": "enrich_excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "excel file",
                        "name": "excel_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/excel/products/errors": {
            "post": {
                "description": "download excel table template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excel"
                ],
                "summary": "ExcelProductsWithErrors",
                "operationId": "excel_products_with_errors",
                "parameters": [
                    {
                        "description": "The input body.\u003cbr /\u003e UploadID is ID of previously uploaded task.",
                        "name": "InputBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadIdRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errinfo.errInf"
                        }
                    }
                }
            }
        },
        "/api/v1/product_history": {
            "post": {
                "description": "get product list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetProductsHistory",
                "operationId": "products_history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit of contents",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "offset of contents",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The input body.\u003cbr /\u003e UploadID is ID of previously uploaded task.",
                        "name": "InputBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errinfo.errInf"
                        }
                    }
                }
            }
        },
        "/api/v1/task_history": {
            "get": {
                "description": "get task list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetSupplierTaskHistory",
                "operationId": "supplier_task_history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit of contents",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "offset of contents",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of user",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of supplier",
                        "name": "X-Supplier-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TaskPublic"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errinfo.errInf"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errinfo.errInf": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "articleCriteria": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ArticleCriteria"
                    }
                },
                "articleNumber": {
                    "type": "string"
                },
                "crossNumbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CrossNumbers"
                    }
                },
                "genericArticleDescription": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "linkageTargets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.LinkageTargets"
                    }
                },
                "mfrName": {
                    "type": "string"
                },
                "oemNumbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OEM"
                    }
                },
                "packageArticleCriteria": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ArticleCriteria"
                    }
                }
            }
        },
        "model.ArticleCriteria": {
            "type": "object",
            "properties": {
                "criteriaAbbrDescription": {
                    "type": "string"
                },
                "criteriaDescription": {
                    "type": "string"
                },
                "criteriaType": {
                    "type": "string"
                },
                "criteriaUnitDescription": {
                    "type": "string"
                },
                "formattedValue": {
                    "type": "string"
                },
                "rawValue": {
                    "type": "string"
                }
            }
        },
        "model.CrossNumbers": {
            "type": "object",
            "properties": {
                "articleNumber": {
                    "type": "string"
                },
                "mfrName": {
                    "type": "string"
                }
            }
        },
        "model.GetTecDocArticlesRequest": {
            "type": "object",
            "properties": {
                "articleNumber": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                }
            }
        },
        "model.LinkageTargets": {
            "type": "object",
            "properties": {
                "beginYearMonth": {
                    "type": "string"
                },
                "endYearMonth": {
                    "type": "string"
                },
                "linkageTargetId": {
                    "type": "integer"
                },
                "mfrName": {
                    "type": "string"
                },
                "vehicleModelSeriesName": {
                    "type": "string"
                }
            }
        },
        "model.OEM": {
            "type": "object",
            "properties": {
                "articleNumber": {
                    "type": "string"
                },
                "mfrName": {
                    "type": "string"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "default": 1
                },
                "article": {
                    "type": "string"
                },
                "articleSupplier": {
                    "type": "string"
                },
                "barcode": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "errorResponse": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "subject": {
                    "type": "string"
                },
                "updateDate": {
                    "type": "string"
                },
                "uploadDate": {
                    "type": "string"
                },
                "uploadId": {
                    "type": "integer"
                }
            }
        },
        "model.TaskPublic": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "productsFailed": {
                    "type": "integer"
                },
                "productsProcessed": {
                    "type": "integer"
                },
                "productsTotal": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "uploadDate": {
                    "type": "string"
                }
            }
        },
        "model.UploadIdRequest": {
            "type": "object",
            "properties": {
                "uploadID": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}